+++ src/components/modal/AlertModal.tsx
@@ -1,7 +1,7 @@
 'use client';
 
 import BaseModal from './BaseModal';
-import Button from '../button';
+import Button from '../ui/Button';
 
 interface Props {
   open: boolean;
@@ -25,13 +25,13 @@ function AlertModal({
       {/* 텍스트 영역 */}
       <div className="text-center space-y-3">
         {/* 타이틀 */}
-        <h2 className='text-xl font-semibold text-basic-black'>{title}</h2>
+        <h2 className="text-xl font-semibold text-basic-black">{title}</h2>
 
         {/* 설명 */}
-        <p className='text-sm font-normal text-gray-04'>{description}</p>
+        <p className="text-sm font-normal text-gray-04">{description}</p>
 
         {/* 세부 사항 */}
-        <p className='text-base font-medium text-gray-08'>{detail}</p>
+        <p className="text-base font-medium text-gray-08">{detail}</p>
       </div>
 
       {/* 버튼 영역 */}
